# feat(core): Add DOM & EventTarget