--- a/NativeScript.code-workspace
+++ b/NativeScript.code-workspace
@@ -1,0 +1,13 @@
+{
+    "folders": [
+        {
+            "path": "."
+        },
+        {
+            "path": "../dominative"
+        },
+        {
+            "path": "../../apps/undom-ng"
+        }
+    ]
+}