--- a/apps/toolbox/src/main.ts
+++ b/apps/toolbox/src/main.ts
@@ -1,3 +1,38 @@
-import { Application, Utils } from '@nativescript/core';
+import { Application, Label, View, FlexboxLayout } from '@nativescript/core';
 
-Application.run({ moduleName: 'app-root' });
+const bench = (fn: Function, times = 10) => {
+	console.time(`bench: ${fn.name}`);
+	for (let i = 0; i < times; i++) {
+		fn();
+	}
+	console.timeEnd(`bench: ${fn.name}`);
+};
+
+// //@ts-ignore
+// registerElement('view', FlexboxLayout);
+
+// declare global {
+// 	interface HTMLElementTagNameMap {
+// 		view: View;
+// 	}
+// }
+
+Application.run({
+	create: () => {
+		// bench(function warmup() {
+		// 	new FlexboxLayout();
+		// }, 1000);
+		const element = new FlexboxLayout();
+		bench(() => {
+			bench(function addEvent() {
+				element.addEventListener('loaded', (data) => {
+					console.log(data);
+				});
+			}, 1);
+		}, 100);
+
+		const view = new FlexboxLayout();
+
+		return view;
+	},
+});