--- a/packages/core/index.ts
+++ b/packages/core/index.ts
@@ -7,7 +7,7 @@ export type { ApplicationEventData, LaunchEventData, OrientationChangedEventData
 
 import { fontScaleChangedEvent, launchEvent, displayedEvent, uncaughtErrorEvent, discardedErrorEvent, suspendEvent, resumeEvent, exitEvent, lowMemoryEvent, orientationChangedEvent, systemAppearanceChanged, systemAppearanceChangedEvent, getMainEntry, getRootView, _resetRootView, getResources, setResources, setCssFileName, getCssFileName, loadAppCss, addCss, on, off, notify, hasListeners, run, orientation, getNativeApplication, hasLaunched, android as appAndroid, ios as iosApp, systemAppearance, setAutoSystemAppearanceChanged, ensureNativeApplication, setMaxRefreshRate } from './application';
 import { inBackground, suspended } from './application/application-common';
-export { Window } from './ui/core/dom/index';
+export * from './ui/core/dom/index';
 
 export const Application = {
 	launchEvent,