--- a/packages/core/ui/core/dom/index.ts
+++ b/packages/core/ui/core/dom/index.ts
@@ -1,0 +1,5 @@
+export { HTMLLabelElement } from './src/nodes/html-label-element/HTMLLabelELement';
+import GlobalWindow from './src/window/Window';
+
+export const Window = GlobalWindow;
+new Window().bindToGlobal();