--- a/packages/core/ui/core/dom/src/config/ChildLessElements.ts
+++ b/packages/core/ui/core/dom/src/config/ChildLessElements.ts
@@ -1,0 +1,1 @@
+export default ['style', 'script'];