--- a/packages/core/ui/core/dom/src/config/NamespaceURI.ts
+++ b/packages/core/ui/core/dom/src/config/NamespaceURI.ts
@@ -1,0 +1,5 @@
+export default {
+	html: 'http://www.w3.org/1999/xhtml',
+	svg: 'http://www.w3.org/2000/svg',
+	mathML: 'http://www.w3.org/1998/Math/MathML',
+};