--- a/packages/core/ui/core/dom/src/config/UnnestableElements.ts
+++ b/packages/core/ui/core/dom/src/config/UnnestableElements.ts
@@ -1,0 +1,1 @@
+export default ['a', 'button', 'dd', 'dt', 'form', 'h1', 'h2', 'h3', 'h4', 'h5', 'h6', 'li', 'option', 'p', 'select', 'table'];