--- a/packages/core/ui/core/dom/src/config/VoidElements.ts
+++ b/packages/core/ui/core/dom/src/config/VoidElements.ts
@@ -1,0 +1,1 @@
+export default ['area', 'base', 'br', 'col', 'embed', 'hr', 'img', 'input', 'link', 'meta', 'param', 'source', 'track', 'wbr', 'meta'];