… apps/automated/package.json
… apps/toolbox/package.json
… apps/toolbox/tsconfig.json
… package.json
… packages/core/package.json
… packages/core/ui/core/dom/src/config/ElementTag.ts
… packages/core/ui/core/dom/src/config/NonImplemenetedElementClasses.ts
… packages/core/ui/core/dom/src/cssstylesheet/CSSStyleSheet.ts
… packages/core/ui/core/dom/src/custom-element/CustomElementRegistry.ts
… packages/core/ui/core/dom/src/dom-implementation/DOMImplementation.ts
… packages/core/ui/core/dom/src/event/AbortSignal.ts
… packages/core/ui/core/dom/src/event/CustomEvent.ts
… packages/core/ui/core/dom/src/event/Event.ts
… packages/core/ui/core/dom/src/nodes/character-data/CharacterData.ts
… packages/core/ui/core/dom/src/nodes/comment/Comment.ts
… packages/core/ui/core/dom/src/nodes/document-fragment/DocumentFragment.ts
… packages/core/ui/core/dom/src/nodes/document-type/DocumentType.ts
… packages/core/ui/core/dom/src/nodes/document/Document.ts
… packages/core/ui/core/dom/src/nodes/element/DOMTokenList.ts
… packages/core/ui/core/dom/src/nodes/element/Element.ts
… packages/core/ui/core/dom/src/nodes/html-element/HTMLElement.ts
… packages/core/ui/core/dom/src/nodes/html-item-template-element/HTMLItemTemplateElement.ts
… packages/core/ui/core/dom/src/nodes/html-label-element/HTMLLabelELement.ts
… packages/core/ui/core/dom/src/nodes/html-prop-element/HTMLPropElement.ts
… packages/core/ui/core/dom/src/nodes/html-slot-element/HTMLSlotElement.ts
… packages/core/ui/core/dom/src/nodes/html-template-element/HTMLTemplateElement.ts
… packages/core/ui/core/dom/src/nodes/node/Node.ts
… packages/core/ui/core/dom/src/nodes/node/NodeList.ts
… packages/core/ui/core/dom/src/nodes/node/NodeTypeEnum.ts
… packages/core/ui/core/dom/src/nodes/parent-node/ParentNode.ts