This is my humble attempt to conform core to DOM & `EventTarget`.

Here's what's added:

- [x] Node
- [x] ParentNode
- [x] Element
- [x] HTMLElement
- [x] Text
- [x] Comment
- [x] CharacterData
- [x] Event
- [x] CustomEvent
- [x] All base classes implement DOM API & Extend `HTMLElement` & `EventTarget`
- [x] EventTarget: `Observable` now conforms to `EventTarget` without breaking current observable implementation.
- [x] Document
- [x] DocumentFragment
- [x] Window
- [x] Key/Array Prop elements
- [x] ItemTemplate element
- [x] Shadow Root
- [x] ShadowHost
- [x] Shadow DOM
- [x] Slot
- [x] Template
- [x] DocumentType
- [x] XmlParser
- [x] XmlSerializer
- [x] getElementsByTag
- [x] getElementById
- [x] getElementsByClassName
- [x] querySelectorAll
- [x] querySelector
- [x] CustomElementsRegistry
- [x] connectedCallback
- [x] disconnectedCallback
- [x] attributeChangedCallback/observedAttributes
- [x] CSSStyleSheet
- [x] DOMTokenList
 
## Working web frameworks with minimum configuration:
Most of these will be based on exisiting implementations with dominative.
| Web Framework   |      StackBlitz      | 
|----------|:-------------:|
| [Lit](https://lit.dev/) |  [Example](https://stackblitz.com/edit/nativescript-stackblitz-templates-8p8bv9?file=app%2Flit%2Fwindow.ts,app%2Flit%2Fapp.ts,app%2Flit%2Fns-lit-element.ts&title=NativeScript%20Starter%20TypeScript) | 
| [SolidJS](https://github.com/solidjs/) |    [Example](https://stackblitz.com/edit/nativescript-stackblitz-templates-mgtc8m?file=app%2Fsolidjs%2Fexample.jsx,package.json&title=NativeScript%20Starter%20TypeScript)   |   

Other APIs to add
- Shadow DOM. Will allow us to write platform agnostic UI with [lit](https://lit.dev/) & css. Might also enable us to use MUI3 for [web](https://m3.material.io/develop/web) or a fork of it.
- Prop, Key & Template elements from [dominative](github.com/SudoMaker/domiNATIVE) to simplify handling of lists in different frameworks.

This is largely based on [undom-ng](github.com/ClassicOldSong/undom-ng/) & [dominative](github.com/SudoMaker/dominative) by @ClassicOldSong and [happy-dom](github.com/capricorn86/happy-dom).

This is more about conformance than performance. Although i have tried to keep everything simple and lean and quite happy about the results.

The Observable changes by being backwards compatible result in a some overhead of course which is fixable by simply moving to `dispatchEvent` instead of `notify` which is mostly because values are first set on `EventData` then they get defined on the DOM `Event` again.

Class initialization/Adding & removing listeners has same performance compared to current core. I'll add a benchmark comparison table soon.

The ultimate end-developer using a web framework will be exposed to DOM types from [@nativescript-dom/type](https://github.com/nativescript-dom/types/ ) which we can add in core eventually.

## Benchmarks

### Without DOM:

```
# 1

  JS: new FlexboxLayout() x 39,037 ops/sec ±4.34% (37 runs sampled)
  JS: view.notify (20 Listeners) x 1,110,038 ops/sec ±2.42% (60 runs sampled)
  JS: view.addChild (shallow) x 860 ops/sec ±8.01% (42 runs sampled)
  JS: view.addChild (deep) x 263 ops/sec ±31.06% (46 runs sampled)

# 2

  JS: new FlexboxLayout() x 41,885 ops/sec ±5.71% (39 runs sampled)
  JS: view.notify (20 Listeners) x 1,083,994 ops/sec ±5.60% (56 runs sampled)
  JS: view.addChild (shallow) x 870 ops/sec ±7.84% (42 runs sampled)
  JS: view.addChild (deep) x 258 ops/sec ±33.50% (47 runs sampled)
```

 ### With DOM:

```
# 1

  JS: document.createElement x 31,484 ops/sec ±53.77% (32 runs sampled)
  JS: new View() x 25,338 ops/sec ±90.88% (46 runs sampled)
  JS: element.dispatchEvent (20 Listeners) x 1,005,212 ops/sec ±2.06% (58 runs sampled)
  JS: view.notify (20 Listeners) x 600,275 ops/sec ±1.24% (57 runs sampled)
  JS: element.appendChild (shallow) x 841 ops/sec ±7.96% (43 runs sampled)
  JS: view.addChild (shallow) x 880 ops/sec ±9.67% (36 runs sampled)
  JS: element.appendChild (deep) x 155 ops/sec ±100.27% (38 runs sampled)
  JS: view.addChild (deep) x 316 ops/sec ±3.38% (54 runs sampled)

# 2 reversed sequence of calls

  JS: new View() x 33,129 ops/sec ±48.56% (34 runs sampled)
  JS: document.createElement x 25,208 ops/sec ±97.78% (48 runs sampled)
  JS: view.notify (20 Listeners) x 621,300 ops/sec ±1.21% (56 runs sampled)
  JS: element.dispatchEvent (20 Listeners) x 1,079,043 ops/sec ±1.49% (55 runs sampled)
  JS: view.addChild (shallow) x 881 ops/sec ±6.82% (45 runs sampled)
  JS: element.appendChild (shallow) x 857 ops/sec ±10.77% (34 runs sampled)
  JS: view.addChild (deep) x 303 ops/sec ±3.52% (53 runs sampled)
  JS: element.appendChild (deep) x 308 ops/sec ±8.91% (22 runs sampled)
```

Update: 

After adding some optimizations, domless `notify` and `notify` calls with DOM  perform the same:
```
JS: view.notify (20 Listeners) x 1,068,513 ops/sec ±0.86% (62 runs sampled)
JS: element.dispatchEvent (20 Listeners) x 1,060,945 ops/sec ±1.19% (59 runs sampled)
```


## PR Checklist

- [x] The PR title follows our guidelines: https://github.com/NativeScript/NativeScript/blob/master/tools/notes/CONTRIBUTING.md#commit-messages.
- [ ] There is an issue for the bug/feature this PR is for. To avoid wasting your time, it's best to open a suggestion issue first and wait for approval before working on it.
- [x] You have signed the [CLA](http://www.nativescript.org/cla).
- [x] All existing tests are passing: https://github.com/NativeScript/NativeScript/blob/master/tools/notes/DevelopmentWorkflow.md#running-unit-tests-application.
- [ ] Tests for the changes are included - https://github.com/NativeScript/NativeScript/blob/master/tools/notes/WritingUnitTests.md.

## What is the current behavior?
There is no DOM in core

## What is the new behavior?
There **is** DOM in core

### No breaking changes introduced.