@ammarahm-ed 2023-02-27
All tests pass but I keep seeing this crash:
```
Successfully synced application org.nativescript.UnitTestApp on device emulator-5554.
  JS: Test: --- [TAB-VIEW-NAVIGATION.testBackNavigationToTabViewWithNestedFramesShouldWork] OK, duration: 961.95
  JS: Test: --- [TAB-VIEW-NAVIGATION.testLoadedAndUnloadedAreFired_WhenNavigatingAwayAndBack] OK, duration: 412.85
  JS: Test: --- [TAB-VIEW-NAVIGATION.testWhenNavigatingBackToANonCachedPageContainingATabViewWithAListViewTheListViewIsThere] OK, duration: 641.98
  JS: Test: TAB-VIEW-NAVIGATION COMPLETED for 2018.64 BACKSTACK DEPTH: 1
  JS: Test:
  JS: === ALL TESTS COMPLETE ===
  JS: 3 OK, 0 failed
  JS: DURATION: 2020.16 ms
  JS: === END OF TESTS ===
  JS: Test: Tests EOF!
  JS: image
  JS: Test: START IMAGE TESTS.
  JS: Test: --- [IMAGE.test_DimensionsAreRoundedAfterScale] FAILED: com.tns.NativeScriptException: Calling js method destroyItem failed
  JS: Error: java.lang.IllegalStateException: Cannot detach Fragment attached to a different FragmentManager. Fragment FragmentBase_bundle_73170_28_TabFragmentImplementation{147fe05} (332c632d-3fef-4586-b428-47ee23f29de0 id=0x2 tag=android:viewpager:2:0) is already attached to a FragmentManager., Stack: setAdapterItems(file: src/packages/core/ui/tab-view/index.android.ts:683:23)
  JS:   at onUnloaded(file: src/packages/core/ui/tab-view/index.android.ts:606:7)
  JS:   at (file: src/packages/core/ui/core/view-base/index.ts:546:69)
  JS:   at callFunctionWithSuper(file: src/packages/core/ui/core/view-base/index.ts:535:2)
  JS:   at callUnloaded(file: src/packages/core/ui/core/view-base/index.ts:546:7)
  JS:   at unloadView(file: src/packages/core/ui/core/view-base/index.ts:730:8)
  JS:   at _removeViewCore(file:///data/data/or...
  JS: Test: --- [IMAGE.test_Image_Members] OK, duration: 0.23
  System.err: An uncaught Exception occurred on "main" thread.
  System.err: Calling js method onPause failed
  System.err: TypeError: Cannot read property 'setDrawingCacheEnabled' of null
  System.err:
  System.err: StackTrace:
  System.err: TabFragmentImplementation.loadBitmapFromView(file:///data/data/org.nativescript.UnitTestApp/files/app/bundle.js:73224:14)
  System.err:   at TabFragmentImplementation.onPause(file:///data/data/org.nativescript.UnitTestApp/files/app/bundle.js:73212:42)
  System.err:   at com.tns.Runtime.callJSMethodNative(Native Method)
  System.err:   at com.tns.Runtime.dispatchCallJSMethodNative(Runtime.java:1302)
  System.err:   at com.tns.Runtime.callJSMethodImpl(Runtime.java:1188)
  System.err:   at com.tns.Runtime.callJSMethod(Runtime.java:1175)
  System.err:   at com.tns.Runtime.callJSMethod(Runtime.java:1153)
  System.err:   at com.tns.Runtime.callJSMethod(Runtime.java:1149)
  System.err:   at org.nativescript.widgets.FragmentBase_bundle_73170_28_TabFragmentImplementation.onPause(Unknown Source:15)
  System.err:   at androidx.fragment.app.Fragment.performPause(Fragment.java:3305)
  System.err:   at androidx.fragment.app.FragmentStateManager.pause(FragmentStateManager.java:631)
  System.err:   at androidx.fragment.app.FragmentStateManager.moveToExpectedState(FragmentStateManager.java:293)
  System.err:   at androidx.fragment.app.FragmentStore.moveToExpectedState(FragmentStore.java:113)
  System.err:   at androidx.fragment.app.FragmentManager.moveToState(FragmentManager.java:1433)
  System.err:   at androidx.fragment.app.FragmentManager.dispatchStateChange(FragmentManager.java:2977)
  System.err:   at androidx.fragment.app.FragmentManager.dispatchPause(FragmentManager.java:2913)
  System.err:   at androidx.fragment.app.Fragment.performPause(Fragment.java:3298)
  System.err:   at androidx.fragment.app.FragmentStateManager.pause(FragmentStateManager.java:631)
  System.err:   at androidx.fragment.app.FragmentStateManager.moveToExpectedState(FragmentStateManager.java:293)
  System.err:   at androidx.fragment.app.FragmentManager.executeOpsTogether(FragmentManager.java:1899)
  System.err:   at androidx.fragment.app.FragmentManager.removeRedundantOperationsAndExecute(FragmentManager.java:1817)
  System.err:   at androidx.fragment.app.FragmentManager.execPendingActions(FragmentManager.java:1760)
  System.err:   at androidx.fragment.app.FragmentManager$5.run(FragmentManager.java:547)
  System.err:   at android.os.Handler.handleCallback(Handler.java:938)
  System.err:   at android.os.Handler.dispatchMessage(Handler.java:99)
  System.err:   at android.os.Looper.loopOnce(Looper.java:201)
  System.err:   at android.os.Looper.loop(Looper.java:288)
  System.err:   at android.app.ActivityThread.main(ActivityThread.java:7839)
  System.err:   at java.lang.reflect.Method.invoke(Native Method)
  System.err:   at com.android.internal.os.RuntimeInit$MethodAndArgsCaller.run(RuntimeInit.java:548)
  System.err:   at com.android.internal.os.ZygoteInit.main(ZygoteInit.java:1003)
  JS: NativeScriptError: Error: Calling js method onPause failed
  JS: TypeError: Cannot read property 'setDrawingCacheEnabled' of null
```
It **is** related to the Observable changes. Something very specific probably? If i add a couple of checks, then everything is ok. Basically `this.owner.nativeViewProtected` is null when `onPause` gets called:
https://github.com/NativeScript/NativeScript/blob/e613ecf819035a5e38140384570ea747bb3db0f6/packages/core/ui/tab-view/index.android.ts#L108