@ammarahm-ed 2023-03-02
> I am personally worried about performances hit of this PR.
This PR is not about performance but conformance to "something" especially the DOM. Performance can be further optimized.

> Especially when the change is not really needed in the sense it does not add more to N
I am surpised to hear this after having so many debates over DOM. It solves the flavor integration problem. There has been a lot of debate in the community around adding a DOM and especially around dominative by @ClassicOldSong not a viable solution because a lot of people say `DOM` is very basic for web frameworks and should be a part of NativeScript core instead of living as an extra layer above it which most framework implementations do currently. I think a more pointless thing in core is the XML flavor that should have been outside the core. And if an xml flavor can live in core, I think it's okay for NativeScript core to BE the DOM so that all flavors can function the same. We can have things like Shadow DOM, plugins that we write once and use anywhere with Lit html & much more. It does add a lot and also puts the NativeScript core into the right direction. Because right now it does not conform to anything which is painful.

As far as performance is concerned, It's fast because it's based on dominative by @ClassicOldSong. Event emitting is a little slower than current core if you use `notify` functions as I mentioned above. The DOM implementation conforms to EventTarget 1:1 which has some overhead but if you use `dispatchEvent` directly, that overhead is removed automatically.

This allows us to slowly make `core` like DOM and conform EventTarget in phases in plugins & core itself. It removes a lot of overhead of maintaining each flavor. 

Finally this might in future allow us to run NativeScript apps on web & mobile like react native and flutter do. It's something that is very high in demand from most devs and NativeScript lacks.