@ammarahm-ed 2023-03-05
@ClassicOldSong Benchmarks:

Without DOM:

```
# 1

  JS: new FlexboxLayout() x 39,037 ops/sec ±4.34% (37 runs sampled)
  JS: view.notify (20 Listeners) x 1,110,038 ops/sec ±2.42% (60 runs sampled)
  JS: view.addChild (shallow) x 860 ops/sec ±8.01% (42 runs sampled)
  JS: view.addChild (deep) x 263 ops/sec ±31.06% (46 runs sampled)

# 2

  JS: new FlexboxLayout() x 41,885 ops/sec ±5.71% (39 runs sampled)
  JS: view.notify (20 Listeners) x 1,083,994 ops/sec ±5.60% (56 runs sampled)
  JS: view.addChild (shallow) x 870 ops/sec ±7.84% (42 runs sampled)
  JS: view.addChild (deep) x 258 ops/sec ±33.50% (47 runs sampled)
```

With DOM:

```
# 1

  JS: document.createElement x 31,484 ops/sec ±53.77% (32 runs sampled)
  JS: new View() x 25,338 ops/sec ±90.88% (46 runs sampled)
  JS: element.dispatchEvent (20 Listeners) x 1,005,212 ops/sec ±2.06% (58 runs sampled)
  JS: view.notify (20 Listeners) x 600,275 ops/sec ±1.24% (57 runs sampled)
  JS: element.appendChild (shallow) x 841 ops/sec ±7.96% (43 runs sampled)
  JS: view.addChild (shallow) x 880 ops/sec ±9.67% (36 runs sampled)
  JS: element.appendChild (deep) x 155 ops/sec ±100.27% (38 runs sampled)
  JS: view.addChild (deep) x 316 ops/sec ±3.38% (54 runs sampled)

# 2 reversed sequence of calls

  JS: new View() x 33,129 ops/sec ±48.56% (34 runs sampled)
  JS: document.createElement x 25,208 ops/sec ±97.78% (48 runs sampled)
  JS: view.notify (20 Listeners) x 621,300 ops/sec ±1.21% (56 runs sampled)
  JS: element.dispatchEvent (20 Listeners) x 1,079,043 ops/sec ±1.49% (55 runs sampled)
  JS: view.addChild (shallow) x 881 ops/sec ±6.82% (45 runs sampled)
  JS: element.appendChild (shallow) x 857 ops/sec ±10.77% (34 runs sampled)
  JS: view.addChild (deep) x 303 ops/sec ±3.52% (53 runs sampled)
  JS: element.appendChild (deep) x 308 ops/sec ±8.91% (22 runs sampled)
```
As I mentioned before, using `dispatchEvent` performs many times faster than using `notify` with DOM. The overhead is obvious since `notify` is backward compatible so it has to do more work.

Other than that, core with DOM goes head to head with core without DOM.